Revert "Merge branch 'serialization-changes'"
This reverts commit bebf967b0dff9618080561f72c9fd1f4ab05e095, reversing
changes made to 96772332112f4aa69a8b096be244b26a07074029.

+++ Pip-Boy/HeadPiece.cs
@@ -3,7 +3,7 @@
     public class HeadPiece : Apparrel
     {
         #region Constructors
-        public HeadPiece(string name, float weight, ushort value, Effect[] effects, byte DT, bool powerArmor) : base(name, weight, value, effects, DT, powerArmor) { }
+        public HeadPiece(string name, double weight, ushort value, Effect[] effects, byte DT, bool powerArmor) : base(name, weight, value, effects, DT, powerArmor) { }
 
         public HeadPiece() : base() { }
         #endregion
